third_party/WebKit/LayoutTests/fast/innerHTML/806705-crash.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
var d = document.implementation.createHTMLDocument();
var e = d.body.innerHTML = '<noscript>noscript</noscript>';
assert_equals(d.querySelector('noscript').innerHTML, 'noscript');
});
</script>
